THBuoi5

+++ QuanLyBanHang/Data/HoaDon_ChiTiet.cs
@@ -12,7 +12,7 @@ namespace QuanLyBanHang.Data
         public int ID { get; set; }
         public int HoaDonID { get; set; }
         public int SanPhamID { get; set; }
-        public short SoLuongBan { get; set; }
+        public int SoLuongBan { get; set; }
         public int DonGiaBan { get; set; }
 
         public virtual HoaDon HoaDon { get; set; } = null!;
@@ -26,7 +26,7 @@ namespace QuanLyBanHang.Data
         public int HoaDonID { get; set; }
         public int SanPhamID { get; set; }
         public string TenSanPham { get; set; } // Thêm 
-        public short SoLuongBan { get; set; }
+        public int SoLuongBan { get; set; }
         public int DonGiaBan { get; set; }
         public int ThanhTien { get; set; } // Thêm 
     }
